+++ CHANGELOG.md
@@ -32,6 +32,8 @@
   ([#5954](https://github.com/mitmproxy/mitmproxy/pull/5954), @mhils)
 * the `esc` key can now be used to exit the current view
   ([#6087](https://github.com/mitmproxy/mitmproxy/pull/6087), @sujaldev)
+* focus-follow shortcut will now work in flow view context too.
+  ([#6088](https://github.com/mitmproxy/mitmproxy/pull/6088), @sujaldev)
 
 ### Breaking Changes